List of Instruments

Method to list segments and instruments currently available through the Trading API.

trading_instruments(
  connection,
  request,
  sec_detailed = FALSE,
  market_id = "ROFX",
  segment_id,
  cfi_code,
  sec_type
)

Arguments

connection
S4. Mandaroty Formal rRofexConnection class object

request
String. Mandaroty The type of request that you are making:

segments: List available market segments

securities: List available instruments listed on Matba Rofex. Depends on 'sec_detailed'.

by_segment: List available instruments searching by market segment. Depends on 'market_id' and 'segment_id'

by_cfi_code: List available instruments searching by CFI Code. Depends on 'cfi_code'

by_type: List available instruments searching by Instrument Type. See section Instrument Types. Depends on 'sec_detailed' and 'sec_type'.

sec_detailed
Logical. Optional for request='securities'. Brings aditional information like segment, price, minimal/maximal trading quantity, settlement date, etc.

market_id
String. Needed for request='by_segment'. Market ID.

ROFX: Matba Rofex

segment_id
String. Needed for request='by_segment'. Market Segment ID.

DDF: Financial Derivatives

DDA: Agropecuarial Derivatives

DUAL: Other Derivatives

MERV: S&P Merval

cfi_code
String. Needed for request='by_cfi_code'. CFI Code. See https://www.quotemedia.com/apifeeds/cfi_code

sec_type
String. Needed for request='by_type'.

E: Equities

D: Debt

C: Collective Investment Vehicles

R: Entitlements (Rights)

O: Listed Options

F: Futures

T: Referencial Instruments

M: Others

Value

If correct, it will load a tibble data frame

See also