Market Data Real Time

This method brings Market Data in Real Time.

trading_md(
  connection,
  market_id = "ROFX",
  symbol,
  entries = c("BI", "OF", "LA", "OP", "CL", "SE", "OI", "HI", "LO", "TV", "IV", "EV",
    "NV", "TC"),
  depth = 1L,
  tidy = TRUE
)

Arguments

connection
S4. Mandatory. Formal rRofexConnection class object

market_id
String. Market to which you are going to connect.

ROFX - Matba Rofex

symbol
String. Mandatory. Use trading_instruments to see which symbols are available.

entries
Vector of Strings. It contains the information to be queried:

BI - Bid.

OF - Offer.

LA - Last Available Price.

OP - Open Price.

CL - Close Price.

SE - Settlement Price.

OI - Open Interest.

HI - Trading Session High Price

LO - Trading Session Low Price

TV - Trading Volume

IV - Index Value

EV - Trading Effective Volume

NV - Nominal Volume

TC - Trade Count

depth
Integer. Depth of the book.

tidy
Logical. Data arranged on a tidy format.

Value

If correct, it will load a tibble data frame

See also

Examples

# If you want to query many products at once,
# I recommend you to use "purrr::map" family like this:

if (FALSE) {
purrr::map_df(
list('MERV - XMEV - GGAL - 48hs','MERV - XMEV - BYMA - 48hs'),
~trading_md(connection = conn, symbol = .x, entries = c("LA","OP", "NV"), tidy = T)
)
}