Send Order to the Market

The method trading_new_order is use to send orders.

trading_new_order(
  connection,
  symbol,
  side,
  quantity,
  price,
  order_type = "Limit",
  time_in_force = "Day",
  iceberg = FALSE,
  expire_date = NULL,
  display_quantity = NULL,
  account
)

Arguments

connection
S4. Mandaroty Formal rRofexConnection class object

symbol
String. Use trading_instruments to see which symbols are available.

side
String. Either 'Buy' or 'Sell'.

quantity
Numeric. Quantity of the order.

price
Numeric. Price of the order.

order_type
String. Type of order.

Limit. Limit order sets the maximum or minimum price at which you are willing to buy or sell.

MLL. Market with Leftover as Limit (market order then unexecuted quantity becomes limit order at last price).

time_in_force
String. Specifies how long the order remains in effect. Absence of this field is interpreted as 'Day':

Day. Day or session.

IOC. Immediate or Cancel.

FOK. Fill or Kill.

GTD. Good Till Date.

iceberg
Logical: if TRUE, then the order is 'iceberg'. FALSE as default.

expire_date
String. Only for GDT orders. Maturity date of the order, With format '%Y-%m-%d'.

display_quantity
Numeric. Only for Iceberg orders. Indicate the disclosed quantity for the 'iceberg' order.

account
String. Account Number / Account ID.

Value

List with outputs like state of the order.

See also